got CORS and CSRF working...
Needed to add the @Bean CorsConfigurationSource back into DefKoiRestSecurity.
Also needed to configure the cookie domain to be the parent so defkon could store defkoi cookies.
Found missing config to $.ajax(): xhrFields {withCredentials: true}.
Added semaphore to refresh daemons to prevent multiple calls from spawning multiple daemons.
Keycloak had to be configured to allow PATCH method.
This was not obvious at first because the initial request is not going to Keycloak, but Keycloak must also have a CORS filter.